Include libmyQueue in the CMake. Place declarationand definition of queue in a single hpp file
